--- a/src/globus_sdk/authorizers/refresh_token.py
+++ b/src/globus_sdk/authorizers/refresh_token.py
@@ -1,9 +1,10 @@
 import logging
-from typing import Any, Callable, Dict, Optional
-
-from globus_sdk.services.auth import AuthClient, OAuthTokenResponse
+from typing import TYPE_CHECKING, Any, Callable, Dict, Optional
 
 from .renewing import RenewingAuthorizer
+
+if TYPE_CHECKING:
+    from globus_sdk.services.auth import AuthClient, OAuthTokenResponse
 
 log = logging.getLogger(__name__)
 
@@ -53,11 +54,11 @@ class RefreshTokenAuthorizer(RenewingAuthorizer):
     def __init__(
         self,
         refresh_token: str,
-        auth_client: AuthClient,
+        auth_client: "AuthClient",
         *,
         access_token: Optional[str] = None,
         expires_at: Optional[int] = None,
-        on_refresh: Optional[Callable[[OAuthTokenResponse], Any]] = None,
+        on_refresh: Optional[Callable[["OAuthTokenResponse"], Any]] = None,
     ):
         log.info(
             "Setting up RefreshTokenAuthorizer with auth_client="
@@ -70,13 +71,13 @@ class RefreshTokenAuthorizer(RenewingAuthorizer):
 
         super().__init__(access_token, expires_at, on_refresh)
 
-    def _get_token_response(self) -> OAuthTokenResponse:
+    def _get_token_response(self) -> "OAuthTokenResponse":
         """
         Make a refresh token grant
         """
         return self.auth_client.oauth2_refresh_token(self.refresh_token)
 
-    def _extract_token_data(self, res: OAuthTokenResponse) -> Dict[str, Any]:
+    def _extract_token_data(self, res: "OAuthTokenResponse") -> Dict[str, Any]:
         """
         Get the tokens .by_resource_server,
         Ensure that only one token was gotten, and return that token.