--- a/src/globus_sdk/authorizers/renewing.py
+++ b/src/globus_sdk/authorizers/renewing.py
@@ -1,12 +1,14 @@
 import abc
 import logging
 import time
-from typing import Any, Callable, Dict, Optional
+from typing import TYPE_CHECKING, Any, Callable, Dict, Optional
 
 from globus_sdk import exc, utils
-from globus_sdk.services.auth import OAuthTokenResponse
 
 from .base import GlobusAuthorizer
+
+if TYPE_CHECKING:
+    from globus_sdk.services.auth import OAuthTokenResponse
 
 log = logging.getLogger(__name__)
 # Provides a buffer for token expiration time to account for
@@ -50,7 +52,7 @@ class RenewingAuthorizer(GlobusAuthorizer, metaclass=abc.ABCMeta):
         self,
         access_token: Optional[str] = None,
         expires_at: Optional[int] = None,
-        on_refresh: Optional[Callable[[OAuthTokenResponse], Any]] = None,
+        on_refresh: Optional[Callable[["OAuthTokenResponse"], Any]] = None,
     ):
         self._access_token = None
         self._access_token_hash = None
@@ -96,14 +98,14 @@ class RenewingAuthorizer(GlobusAuthorizer, metaclass=abc.ABCMeta):
             self._access_token_hash = utils.sha256_string(value)
 
     @abc.abstractmethod
-    def _get_token_response(self) -> OAuthTokenResponse:
+    def _get_token_response(self) -> "OAuthTokenResponse":
         """
         Using whatever method the specific authorizer implementing this class
         does, get a new token response.
         """
 
     @abc.abstractmethod
-    def _extract_token_data(self, res: OAuthTokenResponse) -> Dict[str, Any]:
+    def _extract_token_data(self, res: "OAuthTokenResponse") -> Dict[str, Any]:
         """
         Given a token response object, get the first element of
         token_response.by_resource_server